[ID4] Accumulate LoRA bake windows in F32
Accumulate each bounded effective-weight window in F32.

Resident base weights remain compact FP8, and LoRA factors remain BF16.

Decode once and apply every adapter without intermediate BF16 rounding.

Derive each row scale and encode once at the final storage boundary.

Cover composed sub-ULP updates that BF16 intermediates would lose.

+++ experimental/id4/ideogram4/lora_bake_plan.c
@@ -275,7 +275,7 @@ static iree_status_t id4_ideogram4_lora_bake_calculate_working_set(
       !iree_device_size_checked_mul(up_element_count, sizeof(uint16_t),
                                     &up_byte_length) ||
       !iree_device_size_checked_mul(effective_weight_element_count,
-                                    sizeof(uint16_t),
+                                    sizeof(float),
                                     &effective_weight_byte_length) ||
       !iree_device_size_checked_mul(adapter_count, sizeof(float),
                                     &strength_byte_length)) {
